:green_heart: Fix release action (#27)
* :boom: Replaces javascript action with composite action

This commit replaces the javascript-based action with a composite gh action.
This was done since the javascript had several downsides:
 - Debugging unnecessarily complicated since multiple bash commands were executed
 inside subshells.
 - Maintainers needed to remember to build the javascript package before deploying.

This commit contains the following breaking changes:
 - It does no longer support windows. This, however, can be easily added might
 somebody need it.
 - This version depricates the (redundant)`update_pip, install_deps, dev_install,
 req_file_path, flake8_version, exclude, ignore, max_line_length` action input
 arguments.

* :recycle: Changes entrypoint.sh shebang to /bin/bash

* :green_heart: Makes gh-action test python matrix smaller

Github was not really happy with the gh-action agent creating this many
annotations. I therfore now only test the action with the python 3.8. We
might have to update this in the future.

* :green_heart: Removes fail-on-error flag from gh-action test

* :bulb: Adds -fail-on-error bug comment

* :sparkles: Removes reviewdog bin after main action has run

This commit makes sure the reviewdog binary is removed after the main
action has run. It further contains some small bug fixes and
documentation changes.

* :green_heart: Fixes release gh-action